+++ mainProject/vrTraining.py
@@ -3,7 +3,11 @@ import os
 import sys
 import subprocess
 import atexit
-import keyboard
+from screeninfo import get_monitors
+
+if sys.platform == "win32":
+    import keyboard
+
 from direct.gui.DirectGui import *
 from panda3d.core import *
 from panda3d.core import (
@@ -32,7 +36,7 @@ PROGRAM_TUTORIAL_SCRIPT_PATH = os.path.abspath(
 loadPrcFileData(
     "",
     f"""want-pstats 0
-win-size 1920 1080
+win-size {get_monitors()[0].width} {get_monitors()[0].height}
 fullscreen 0
 undecorated 0
 show-frame-rate-meter 1
@@ -60,12 +64,13 @@ class Launcher(BaseVrApp):
             wantVr=False,
         )
         self.setBackgroundColor(0, 0, 0, 1)
-        keyboard.add_word_listener(
-            word="exit",
-            callback=lambda: os.system(f"taskkill /F /PID {os.getpid()}"),
-            triggers=["enter"],
-            timeout=1,
-        )
+        if sys.platform == "win32":
+            keyboard.add_word_listener(
+                word="exit",
+                callback=lambda: os.system(f"taskkill /F /PID {os.getpid()}"),
+                triggers=["enter"],
+                timeout=1,
+            )
 
     def makeCard(self, name):
         cm = CardMaker(name)
@@ -84,11 +89,12 @@ class Launcher(BaseVrApp):
         self.backgroundImageNp.setColor(1, 1, 1, 1)
         self.panel = Panel()
         runOnce = False
-        for tutorialImage in os.listdir("Training/tutorials/VR/"):
+        os.chdir("mainProject")
+        for tutorialImage in os.listdir(os.path.join("training", "tutorials", "VR")):
             if tutorialImage.endswith(".png"):
                 name = os.path.splitext(tutorialImage)[0]
                 texture = self.loader.loadTexture(
-                    os.path.join("Training/tutorials/VR/", tutorialImage)
+                    os.path.join("training", "tutorials", "VR", tutorialImage)
                 )
                 name = name.split("--")
                 self.panel.addPage(
